refactor(watch): split distance/cutoff predicates + log formatters + classifiers into WatchModeController.Diagnostics.cs
